Refactor: Removed redundant test and moved SearchController into correct package.
